requirements.txt
asgiref==3.4.1
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cycler==0.11.0
cymem==2.0.6
Django==4.0
djangorestframework==3.13.1
fonttools==4.28.5
idna==3.3
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
langcodes==3.3.0
MarkupSafe==2.0.1
matplotlib==3.5.1
murmurhash==1.0.6
nltk==3.6.7
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pathy==0.6.1
Pillow==8.4.0
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
regex==2021.11.10
requests==2.26.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
smart-open==5.2.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
sqlparse==0.4.2
srsly==2.4.2
thinc==8.0.13
threadpoolctl==3.0.0
tqdm==4.62.3
typer==0.4.0
typing-extensions==4.0.1
urllib3==1.26.7
wasabi==0.9.0
